# CUDA: refactor and optimize IQ MMVQ