This PR refactors and optimizes the IQ MMVQ CUDA code. Notably as part of these changes I'm changing some values in `ggml-common.h`. The "qr" values are meant to represent how many low bit data values are contained in a single 8 bit integer. This value is used to derive "qi" which represents how many 32 bit integers are needed to represent the low bit data values of a quantized block. These values are intended to be properties of the data type independent of any kernels.

In MMVQ qr and qi are used to determine how one load of 32 integers for the quantized weights needs to be aligned with the loads of the q8 activations. It is oftentimes beneficial to load more values at once which is intended to be done via the "vdr" value which is a factor that increases the number of simultaneous loads so that the total stride per invocation of `vec_dot_q_cuda` is qr*vdr. However, for the IQ quants this was instead done by increasing QR. This does not matter for MMVQ but it's a problem for MMQ where the values of qr and qi matter for determining how much shared memory needs to be allocated and how the activations need to be loaded. So for this reason I'm changing the qr and qi values of the IQ quants to the originally intended values. Notably this affects the SYCL backend but I am **not** able to test the corresponding changes myself due to a lack of Intel hardware. @arthw @airMeng I don't know who to tag in terms of llama.cpp SYCL developers; please either test my changes or tell me who I should contact.

<details>
<summary>Performance changes</summary>

| Model                         | GPU      |   Microbatch size | Test   |   t/s master |   t/s 59f5fe59 |   Speedup |
|:------------------------------|:---------|------------------:|:-------|-------------:|---------------:|----------:|
| llama 8B IQ1_M - 1.75 bpw     | RX 6800  |                 1 | pp512  |        53.40 |          49.22 |      0.92 |
| llama 8B IQ1_M - 1.75 bpw     | RX 6800  |                 2 | pp512  |        85.65 |          86.32 |      1.01 |
| llama 8B IQ1_M - 1.75 bpw     | RX 6800  |                 4 | pp512  |       113.24 |         112.06 |      0.99 |
| llama 8B IQ1_M - 1.75 bpw     | RX 6800  |                 8 | pp512  |       163.77 |         137.24 |      0.84 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 3090 |                 1 | pp512  |       157.19 |         167.03 |      1.06 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 3090 |                 2 | pp512  |       257.59 |         271.74 |      1.05 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 3090 |                 4 | pp512  |       369.37 |         396.47 |      1.07 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 3090 |                 8 | pp512  |       509.12 |         577.37 |      1.13 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 4090 |                 1 | pp512  |       303.46 |         315.04 |      1.04 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 4090 |                 2 | pp512  |       491.38 |         505.95 |      1.03 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 4090 |                 4 | pp512  |       768.49 |         800.65 |      1.04 |
| llama 8B IQ1_M - 1.75 bpw     | RTX 4090 |                 8 | pp512  |      1104.59 |        1199.08 |      1.09 |
| llama 8B IQ1_M - 1.75 bpw     | P40      |                 1 | pp512  |        51.24 |          55.00 |      1.07 |
| llama 8B IQ1_M - 1.75 bpw     | P40      |                 2 | pp512  |        59.62 |          62.14 |      1.04 |
| llama 8B IQ1_M - 1.75 bpw     | P40      |                 4 | pp512  |        79.55 |          87.26 |      1.10 |
| llama 8B IQ1_M - 1.75 bpw     | P40      |                 8 | pp512  |       100.85 |         123.60 |      1.23 |
| llama 8B IQ1_S - 1.5625 bpw   | RX 6800  |                 1 | pp512  |        55.56 |          54.13 |      0.97 |
| llama 8B IQ1_S - 1.5625 bpw   | RX 6800  |                 2 | pp512  |        89.01 |          92.03 |      1.03 |
| llama 8B IQ1_S - 1.5625 bpw   | RX 6800  |                 4 | pp512  |       119.90 |         143.32 |      1.20 |
| llama 8B IQ1_S - 1.5625 bpw   | RX 6800  |                 8 | pp512  |       163.32 |         177.29 |      1.09 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 3090 |                 1 | pp512  |       171.96 |         182.08 |      1.06 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 3090 |                 2 | pp512  |       294.86 |         321.29 |      1.09 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 3090 |                 4 | pp512  |       400.41 |         454.23 |      1.13 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 3090 |                 8 | pp512  |       547.64 |         613.82 |      1.12 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 4090 |                 1 | pp512  |       314.54 |         324.53 |      1.03 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 4090 |                 2 | pp512  |       532.28 |         536.75 |      1.01 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 4090 |                 4 | pp512  |       802.89 |         869.56 |      1.08 |
| llama 8B IQ1_S - 1.5625 bpw   | RTX 4090 |                 8 | pp512  |      1135.76 |        1215.14 |      1.07 |
| llama 8B IQ1_S - 1.5625 bpw   | P40      |                 1 | pp512  |        54.77 |          60.16 |      1.10 |
| llama 8B IQ1_S - 1.5625 bpw   | P40      |                 2 | pp512  |        62.83 |          65.50 |      1.04 |
| llama 8B IQ1_S - 1.5625 bpw   | P40      |                 4 | pp512  |        83.18 |          91.52 |      1.10 |
| llama 8B IQ1_S - 1.5625 bpw   | P40      |                 8 | pp512  |       103.75 |         101.55 |      0.98 |
| llama 8B IQ2_M - 2.7 bpw      | RX 6800  |                 1 | pp512  |        37.80 |          40.21 |      1.06 |
| llama 8B IQ2_M - 2.7 bpw      | RX 6800  |                 2 | pp512  |        65.40 |          71.74 |      1.10 |
| llama 8B IQ2_M - 2.7 bpw      | RX 6800  |                 4 | pp512  |        89.33 |         108.04 |      1.21 |
| llama 8B IQ2_M - 2.7 bpw      | RX 6800  |                 8 | pp512  |       118.42 |         138.37 |      1.17 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 3090 |                 1 | pp512  |       144.67 |         154.45 |      1.07 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 3090 |                 2 | pp512  |       247.07 |         257.73 |      1.04 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 3090 |                 4 | pp512  |       355.66 |         364.56 |      1.03 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 3090 |                 8 | pp512  |       523.70 |         538.26 |      1.03 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 4090 |                 1 | pp512  |       256.04 |         256.11 |      1.00 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 4090 |                 2 | pp512  |       436.39 |         434.25 |      1.00 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 4090 |                 4 | pp512  |       691.87 |         693.52 |      1.00 |
| llama 8B IQ2_M - 2.7 bpw      | RTX 4090 |                 8 | pp512  |      1149.50 |        1154.73 |      1.00 |
| llama 8B IQ2_M - 2.7 bpw      | P40      |                 1 | pp512  |        50.45 |          53.80 |      1.07 |
| llama 8B IQ2_M - 2.7 bpw      | P40      |                 2 | pp512  |        55.44 |          58.77 |      1.06 |
| llama 8B IQ2_M - 2.7 bpw      | P40      |                 4 | pp512  |        80.85 |          83.89 |      1.04 |
| llama 8B IQ2_M - 2.7 bpw      | P40      |                 8 | pp512  |       119.35 |         120.94 |      1.01 |
| llama 8B IQ2_S - 2.5 bpw      | RX 6800  |                 1 | pp512  |        46.08 |          46.31 |      1.00 |
| llama 8B IQ2_S - 2.5 bpw      | RX 6800  |                 2 | pp512  |        75.10 |          81.27 |      1.08 |
| llama 8B IQ2_S - 2.5 bpw      | RX 6800  |                 4 | pp512  |        96.45 |         123.49 |      1.28 |
| llama 8B IQ2_S - 2.5 bpw      | RX 6800  |                 8 | pp512  |       116.69 |         140.50 |      1.20 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 3090 |                 1 | pp512  |       146.70 |         151.47 |      1.03 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 3090 |                 2 | pp512  |       255.82 |         251.69 |      0.98 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 3090 |                 4 | pp512  |       378.84 |         377.85 |      1.00 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 3090 |                 8 | pp512  |       535.84 |         535.17 |      1.00 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 4090 |                 1 | pp512  |       274.49 |         276.89 |      1.01 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 4090 |                 2 | pp512  |       469.03 |         454.41 |      0.97 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 4090 |                 4 | pp512  |       776.19 |         769.60 |      0.99 |
| llama 8B IQ2_S - 2.5 bpw      | RTX 4090 |                 8 | pp512  |      1184.92 |        1168.13 |      0.99 |
| llama 8B IQ2_S - 2.5 bpw      | P40      |                 1 | pp512  |        46.43 |          46.90 |      1.01 |
| llama 8B IQ2_S - 2.5 bpw      | P40      |                 2 | pp512  |        57.69 |          58.24 |      1.01 |
| llama 8B IQ2_S - 2.5 bpw      | P40      |                 4 | pp512  |        81.95 |          86.35 |      1.05 |
| llama 8B IQ2_S - 2.5 bpw      | P40      |                 8 | pp512  |       117.88 |         117.48 |      1.00 |
| llama 8B IQ2_XS - 2.3125 bpw  | RX 6800  |                 1 | pp512  |        48.85 |          47.87 |      0.98 |
| llama 8B IQ2_XS - 2.3125 bpw  | RX 6800  |                 2 | pp512  |        85.05 |          82.40 |      0.97 |
| llama 8B IQ2_XS - 2.3125 bpw  | RX 6800  |                 4 | pp512  |       113.93 |         122.77 |      1.08 |
| llama 8B IQ2_XS - 2.3125 bpw  | RX 6800  |                 8 | pp512  |       142.12 |         137.76 |      0.97 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 3090 |                 1 | pp512  |       150.82 |         154.55 |      1.02 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 3090 |                 2 | pp512  |       261.42 |         256.79 |      0.98 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 3090 |                 4 | pp512  |       387.68 |         382.61 |      0.99 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 3090 |                 8 | pp512  |       539.88 |         536.58 |      0.99 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 4090 |                 1 | pp512  |       283.86 |         285.14 |      1.00 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 4090 |                 2 | pp512  |       483.41 |         468.26 |      0.97 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 4090 |                 4 | pp512  |       799.44 |         793.42 |      0.99 |
| llama 8B IQ2_XS - 2.3125 bpw  | RTX 4090 |                 8 | pp512  |      1181.49 |        1165.17 |      0.99 |
| llama 8B IQ2_XS - 2.3125 bpw  | P40      |                 1 | pp512  |        48.15 |          48.11 |      1.00 |
| llama 8B IQ2_XS - 2.3125 bpw  | P40      |                 2 | pp512  |        60.15 |          60.02 |      1.00 |
| llama 8B IQ2_XS - 2.3125 bpw  | P40      |                 4 | pp512  |        83.81 |          88.44 |      1.06 |
| llama 8B IQ2_XS - 2.3125 bpw  | P40      |                 8 | pp512  |       119.43 |         117.26 |      0.98 |
| llama 8B IQ2_XXS - 2.0625 bpw | RX 6800  |                 1 | pp512  |        51.04 |          42.05 |      0.82 |
| llama 8B IQ2_XXS - 2.0625 bpw | RX 6800  |                 2 | pp512  |        70.78 |          76.72 |      1.08 |
| llama 8B IQ2_XXS - 2.0625 bpw | RX 6800  |                 4 | pp512  |        90.88 |         122.41 |      1.35 |
| llama 8B IQ2_XXS - 2.0625 bpw | RX 6800  |                 8 | pp512  |       108.26 |         147.99 |      1.37 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 3090 |                 1 | pp512  |       108.48 |         160.28 |      1.48 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 3090 |                 2 | pp512  |       173.38 |         269.81 |      1.56 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 3090 |                 4 | pp512  |       203.23 |         381.82 |      1.88 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 3090 |                 8 | pp512  |       258.22 |         558.16 |      2.16 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 4090 |                 1 | pp512  |       238.67 |         303.88 |      1.27 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 4090 |                 2 | pp512  |       378.23 |         494.32 |      1.31 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 4090 |                 4 | pp512  |       470.38 |         723.03 |      1.54 |
| llama 8B IQ2_XXS - 2.0625 bpw | RTX 4090 |                 8 | pp512  |       633.18 |        1195.46 |      1.89 |
| llama 8B IQ2_XXS - 2.0625 bpw | P40      |                 1 | pp512  |        29.94 |          46.71 |      1.56 |
| llama 8B IQ2_XXS - 2.0625 bpw | P40      |                 2 | pp512  |        40.33 |          54.33 |      1.35 |
| llama 8B IQ2_XXS - 2.0625 bpw | P40      |                 4 | pp512  |        40.90 |          81.67 |      2.00 |
| llama 8B IQ2_XXS - 2.0625 bpw | P40      |                 8 | pp512  |        37.94 |         116.89 |      3.08 |
| llama 8B IQ3_S - 3.4375 bpw   | RX 6800  |                 1 | pp512  |        33.10 |          38.16 |      1.15 |
| llama 8B IQ3_S - 3.4375 bpw   | RX 6800  |                 2 | pp512  |        40.60 |          74.50 |      1.83 |
| llama 8B IQ3_S - 3.4375 bpw   | RX 6800  |                 4 | pp512  |        45.80 |         123.68 |      2.70 |
| llama 8B IQ3_S - 3.4375 bpw   | RX 6800  |                 8 | pp512  |        48.94 |         151.62 |      3.10 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 3090 |                 1 | pp512  |       131.03 |         143.17 |      1.09 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 3090 |                 2 | pp512  |       221.41 |         234.53 |      1.06 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 3090 |                 4 | pp512  |       332.33 |         350.82 |      1.06 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 3090 |                 8 | pp512  |       501.54 |         526.60 |      1.05 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 4090 |                 1 | pp512  |       220.21 |         221.34 |      1.01 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 4090 |                 2 | pp512  |       368.67 |         369.48 |      1.00 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 4090 |                 4 | pp512  |       664.61 |         662.90 |      1.00 |
| llama 8B IQ3_S - 3.4375 bpw   | RTX 4090 |                 8 | pp512  |      1122.96 |        1144.50 |      1.02 |
| llama 8B IQ3_S - 3.4375 bpw   | P40      |                 1 | pp512  |        39.70 |          44.13 |      1.11 |
| llama 8B IQ3_S - 3.4375 bpw   | P40      |                 2 | pp512  |        47.98 |          52.10 |      1.09 |
| llama 8B IQ3_S - 3.4375 bpw   | P40      |                 4 | pp512  |        74.74 |          79.60 |      1.07 |
| llama 8B IQ3_S - 3.4375 bpw   | P40      |                 8 | pp512  |       111.09 |         115.68 |      1.04 |
| llama 8B IQ3_S mix - 3.66 bpw | RX 6800  |                 1 | pp512  |        34.07 |          38.58 |      1.13 |
| llama 8B IQ3_S mix - 3.66 bpw | RX 6800  |                 2 | pp512  |        42.81 |          73.89 |      1.73 |
| llama 8B IQ3_S mix - 3.66 bpw | RX 6800  |                 4 | pp512  |        48.75 |         118.86 |      2.44 |
| llama 8B IQ3_S mix - 3.66 bpw | RX 6800  |                 8 | pp512  |        52.12 |         142.37 |      2.73 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 3090 |                 1 | pp512  |       133.40 |         144.93 |      1.09 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 3090 |                 2 | pp512  |       223.57 |         235.91 |      1.06 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 3090 |                 4 | pp512  |       331.13 |         347.12 |      1.05 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 3090 |                 8 | pp512  |       485.86 |         504.17 |      1.04 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 4090 |                 1 | pp512  |       216.78 |         217.79 |      1.00 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 4090 |                 2 | pp512  |       365.57 |         368.64 |      1.01 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 4090 |                 4 | pp512  |       661.20 |         663.19 |      1.00 |
| llama 8B IQ3_S mix - 3.66 bpw | RTX 4090 |                 8 | pp512  |      1089.92 |        1113.22 |      1.02 |
| llama 8B IQ3_S mix - 3.66 bpw | P40      |                 1 | pp512  |        40.91 |          45.28 |      1.11 |
| llama 8B IQ3_S mix - 3.66 bpw | P40      |                 2 | pp512  |        49.27 |          53.23 |      1.08 |
| llama 8B IQ3_S mix - 3.66 bpw | P40      |                 4 | pp512  |        76.73 |          81.46 |      1.06 |
| llama 8B IQ3_S mix - 3.66 bpw | P40      |                 8 | pp512  |       111.40 |         115.70 |      1.04 |
| llama 8B IQ3_XS - 3.3 bpw     | RX 6800  |                 1 | pp512  |        39.95 |          42.88 |      1.07 |
| llama 8B IQ3_XS - 3.3 bpw     | RX 6800  |                 2 | pp512  |        56.50 |          80.17 |      1.42 |
| llama 8B IQ3_XS - 3.3 bpw     | RX 6800  |                 4 | pp512  |        69.21 |         127.19 |      1.84 |
| llama 8B IQ3_XS - 3.3 bpw     | RX 6800  |                 8 | pp512  |        75.96 |         154.91 |      2.04 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 3090 |                 1 | pp512  |       131.71 |         142.08 |      1.08 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 3090 |                 2 | pp512  |       225.58 |         236.88 |      1.05 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 3090 |                 4 | pp512  |       348.83 |         360.71 |      1.03 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 3090 |                 8 | pp512  |       511.77 |         527.72 |      1.03 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 4090 |                 1 | pp512  |       229.77 |         231.39 |      1.01 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 4090 |                 2 | pp512  |       391.05 |         392.01 |      1.00 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 4090 |                 4 | pp512  |       710.10 |         714.86 |      1.01 |
| llama 8B IQ3_XS - 3.3 bpw     | RTX 4090 |                 8 | pp512  |      1141.69 |        1152.60 |      1.01 |
| llama 8B IQ3_XS - 3.3 bpw     | P40      |                 1 | pp512  |        39.44 |          42.72 |      1.08 |
| llama 8B IQ3_XS - 3.3 bpw     | P40      |                 2 | pp512  |        50.67 |          50.65 |      1.00 |
| llama 8B IQ3_XS - 3.3 bpw     | P40      |                 4 | pp512  |        73.72 |          80.02 |      1.09 |
| llama 8B IQ3_XS - 3.3 bpw     | P40      |                 8 | pp512  |       110.53 |         112.94 |      1.02 |
| llama 8B IQ3_XXS - 3.0625 bpw | RX 6800  |                 1 | pp512  |        46.62 |          46.60 |      1.00 |
| llama 8B IQ3_XXS - 3.0625 bpw | RX 6800  |                 2 | pp512  |        78.05 |          83.07 |      1.06 |
| llama 8B IQ3_XXS - 3.0625 bpw | RX 6800  |                 4 | pp512  |       110.96 |         126.46 |      1.14 |
| llama 8B IQ3_XXS - 3.0625 bpw | RX 6800  |                 8 | pp512  |       131.35 |         155.94 |      1.19 |
| llama 8B IQ3_XXS - 3.0625 bpw | RTX 3090 |                 1 | pp512  |       135.27 |         143.46 |      1.06 |
| llama 8B IQ3_XXS - 3.0625 bpw | RTX 3090 |                 2 | pp512  |       233.96 |         242.70 |      1.04 |
| llama 8B IQ3_XXS - 3.0625 bpw | RTX 3090 |                 4 | pp512  |       360.31 |         367.51 |      1.02 |
| llama 8B IQ3_XXS - 3.0625 bpw | RTX 3090 |                 8 | pp512  |       519.38 |         527.16 |      1.01 |
…