@airMeng 2024-06-30
Thank you for tagging me. I'm not quite familiar with IQ_XXX so I tried ```Llama-3-Lumimaid-8B-v0.1-OAS-IQ4_XS-imat.gguf``` and this PR **works fine**. Is the model one of the targets of this PR? After all the code changes looks good for me.

@OuadiElfarouki @luoyu-intel for awareness.