@JohannesGaessler 2024-06-30
>Is the model one of the targets of this PR? After all the code changes looks good for me.

iq4_xs as well as all iq2 and iq3 models should be affected. Overall the changes I made to SYCL should be very simple, I just can't test whether they actually work.