@slaren 2024-06-30
Shouldn't the removed code allow the IQ quants to work with GPUs without dp4a? I understand that due to the CC check `ggml_cuda_mul_mat` this code was actually never used, but if that was fixed it would allow the IQ quants to work rather than crash on old GPUs.