@JohannesGaessler 2024-06-30
If `__dp4a` is unavailable it would be better to use DMMV. The performance of casting 32 bit integers to arrays of 8 bit integers and doing regular arithmetic with them is quite poor.