@slaren 2024-06-30
Right, but the problem is that dmmv does not support IQ quants.