@JohannesGaessler 2024-06-30
How about this: replace `__dp4a` with `ggml_cuda_dp4a` for which you can then implement a workaround if the instruction is unavailable.