@JohannesGaessler 2024-06-30
Performance with a `__dp4a` workaround was much better than I expected (~90% of regular speed on an RTX 3090) so I made MMVQ the default regardless of compute capability; I'll get someone to test this on a GPU for which the default change would actually make a difference.