@the-crypt-keeper 2024-07-01
# Meta-Llama-3-8B-Instruct-Q6_K.gguf on main
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: Tesla P100-PCIE-16GB, compute capability 6.0, VMM: yes
| model                          |       size |     params | backend    | ngl |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: |
| llama 8B Q6_K                  |   6.14 GiB |     8.03 B | CUDA       |  99 |         pp512 |    511.07 ± 0.00 |
| llama 8B Q6_K                  |   6.14 GiB |     8.03 B | CUDA       |  99 |         tg128 |     16.01 ± 0.00 |

build: 0ddeff10 (3273)

# Meta-Llama-3-8B-Instruct-Q6_K.gguf on branch
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: Tesla P100-PCIE-16GB, compute capability 6.0, VMM: yes
| model                          |       size |     params | backend    | ngl |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: |
| llama 8B Q6_K                  |   6.14 GiB |     8.03 B | CUDA       |  99 |         pp512 |    512.43 ± 0.00 |
| llama 8B Q6_K                  |   6.14 GiB |     8.03 B | CUDA       |  99 |         tg128 |     27.05 ± 0.00 |

build: 30f85eba (3271)

🚀