@the-crypt-keeper 2024-07-01
This PR improved batch performance by almost an order of magnitude at both 2- and 3- streams on my P100 as well:

# Meta-Llama-3-8B-Instruct-Q8_0.gguf batch main

main: n_kv_max = 4096, n_batch = 512, n_ubatch = 512, flash_attn = 1, is_pp_shared = 0, n_gpu_layers = 99, n_threads = 14, n_threads_batch = 14

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.986 |   519.24 |    8.690 |    14.73 |    9.676 |    66.14 |
|   512 |    128 |    2 |   1280 |    1.970 |   519.86 |   20.093 |    12.74 |   22.063 |    58.02 |
|   512 |    128 |    3 |   1920 |    2.975 |   516.30 |   20.679 |    18.57 |   23.654 |    81.17 |
|   512 |    128 |    4 |   2560 |    3.995 |   512.66 |   21.081 |    24.29 |   25.075 |   102.09 |

(not sure why b=2 got WORSE here but I re-ran several times)

# Meta-Llama-3-8B-Instruct-Q8_0.gguf batch branch

main: n_kv_max = 4096, n_batch = 512, n_ubatch = 512, flash_attn = 1, is_pp_shared = 0, n_gpu_layers = 99, n_threads = 14, n_threads_batch = 14

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.987 |   518.75 |    3.575 |    35.81 |    4.562 |   140.29 |
|   512 |    128 |    2 |   1280 |    1.969 |   520.15 |    3.765 |    68.00 |    5.733 |   223.25 |
|   512 |    128 |    3 |   1920 |    2.970 |   517.10 |    4.459 |    86.12 |    7.430 |   258.43 |
|   512 |    128 |    4 |   2560 |    3.991 |   513.13 |    6.111 |    83.78 |   10.102 |   253.41 |

b=2 is now double and b=3 is the sweet spot for overall throughput