@duaneking 2024-07-01
I have a device that CUDA lists as `Device: cuda:0 NVIDIA GeForce GTX 1070 : native` here in my home lab.  Would it be useful?