@smcnally 2024-07-04
Understood cuda-iq-opt-3 is merged. Is test data from a "Maxwell 2.0" / Compute 5.2 GPU still helpful? **llama-simple** and **llama-benchmark-matmult** worked with a Tesla M40. I ran through these before the M40 reached 90oC:

* codellama-7b-instruct.Q5_K_M.gguf
* L3-8B-Lunaris-v1.i1-Q5_K_M.gguf
* Hermes-2-Pro-Llama-3-Instruct-Merged-DPO-Q6_K.gguf
* alphamonarch-7b.Q6_K.gguf
* capybarahermes-2.5-mistral-7b.Q5_K_M.gguf (a few variations with this one)

> `./llama-simple -n 128 -c 4096 -ngl 99 -m /mnt/models/gguf/capybarahermes-2.5-mistral-7b.Q5_K_M.gguf -p "List 5 benefits NVIDIA's Maxwell 2.0 architecture has over Maxwell 1.0. Step through the major benefits CUDA Compute Capability 5.2 has over 5.0."`

**llama-bench**, -batched-bench, -server, and -cli all core dumped. *I have detailed files.*