@smcnally 2024-07-04
Bench and server in ggerganov:master are working great with mixed 5.2 and 6.1


```
./llama-bench -m llava-v1.6-vicuna-13b.Q4_K_M.gguf 
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 3 CUDA devices:
  Device 0: NVIDIA P104-100, compute capability 6.1, VMM: yes
  Device 1: NVIDIA P106-100, compute capability 6.1, VMM: yes
  Device 2: Tesla M40, compute capability 5.2, VMM: yes
| model                          |       size |     params | backend    | ngl |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: |
| llama 13B Q4_K - Medium        |   7.33 GiB |    13.02 B | CUDA       |  99 |         pp512 |    193.24 ± 0.85 |
| llama 13B Q4_K - Medium        |   7.33 GiB |    13.02 B | CUDA       |  99 |         tg128 |     16.67 ± 0.31 |

build: f6190247 (3291)
```
> 
> **llama-bench**, -batched-bench, -server, and -cli all core dumped. 

These are all working in master