@JohannesGaessler 2024-07-04
>Understood cuda-iq-opt-3 is merged. Is test data from a "Maxwell 2.0" / Compute 5.2 GPU still helpful?

More data is always helpful. If it turns out that some changes in this PR were bad they can potentially be reverted.

>llama-bench, -batched-bench, -server, and -cli all core dumped. I have detailed files.

>They are all working in master.

Just to be clear, do you mean that they work on master prior to or after this PR?